Rename Astronoby::RiseTransitSet into Astronoby::Events::ObservationEvents
