+++ lib/astronoby.rb
@@ -11,13 +11,13 @@ require "astronoby/coordinates/horizontal"
 require "astronoby/aberration"
 require "astronoby/equinox_solstice"
 require "astronoby/errors"
+require "astronoby/events/observation_events"
 require "astronoby/geocentric_parallax"
 require "astronoby/mean_obliquity"
 require "astronoby/nutation"
 require "astronoby/observer"
 require "astronoby/precession"
 require "astronoby/refraction"
-require "astronoby/rise_transit_set"
 require "astronoby/time/greenwich_sidereal_time"
 require "astronoby/time/local_sidereal_time"
 require "astronoby/util/astrodynamics"
